+++ src/features/post/api/getHotPosts.ts
@@ -1,10 +1,10 @@
 import { API_SUB_URLS_V3 } from '@/constants/apiConfig';
 import axiosInstance from '@/shared/lib/axios';
 import { IApiResponse } from '@/shared/types/ApiResponse';
-import { Posts } from './getPostList';
+import { Post } from '../types/post';
 
 const getHotPosts = async () => {
-  const response = await axiosInstance.get<IApiResponse<Posts[]>>(`${API_SUB_URLS_V3}/posts/top`);
+  const response = await axiosInstance.get<IApiResponse<Post[]>>(`${API_SUB_URLS_V3}/posts/top`);
 
   return response.data.data;
 };
