+++ src/features/post/api/getMyPosts.ts
@@ -1,28 +1,7 @@
 import { API_SUB_URLS } from '@/constants/apiConfig';
 import axiosInstance from '@/shared/lib/axios';
 import { IApiResponse } from '@/shared/types/ApiResponse';
-
-type Category = {
-  categoryId: number;
-  categoryName: string;
-};
-
-type Author = {
-  userId: number;
-  nickname: string;
-  imgUrl: string;
-};
-
-type Post = {
-  postId: number;
-  problemNumber: number;
-  title: string;
-  createdAt: string; // ISO 문자열 (예: "2025-06-10T00:21:02.664862")
-  categories: Category[];
-  author: Author;
-  commentCount: number;
-  likeCount: number;
-};
+import { Post } from '../types/post';
 
 export interface IGetMyPostsResponse {
   nextCursorId: number;
