+++ src/features/post/types/Category.ts
@@ -1,0 +1,4 @@
+export type Category = {
+  categoryId: number;
+  categoryName: string;
+};
