+++ src/features/post/types/author.ts
@@ -1,0 +1,5 @@
+export type Author = {
+  userId: number;
+  nickname: string;
+  imgUrl: string;
+};
